requirements.txt
flask>=2.3.0
playwright>=1.39.0
aiohttp>=3.8.0
requests>=2.28.0
openai>=1.0.0
pillow>=9.0.0
markdown>=3.4.0
python-dateutil>=2.8.0
jinja2>=3.0.0
beautifulsoup4>=4.11.0
psutil>=5.9.0
apscheduler>=3.10.0
python-dotenv>=1.0.0